+++ src/main/clojure/george/code/paredit.clj
@@ -44,12 +44,29 @@
          parsetree-m
          state-m)))
 
+(defn- correct-pairing
+  "A hack to correct pairing due to strange bug related to after-comment-line.
+  A single paren should never be inserted unpaired at the end of the code.
+  Only correct if at the very end of the text (as this is the only case the bug  seems to appear)."
+  [codearea offset s]
+  ;(println "/correct-paring" "offset:" offset "ca len:" (.getLength codearea))
+  (if (not= offset (.getLength codearea))
+    s
+    (case s
+      "(" "()"
+      ")" "()"
+      "[" "[]"
+      "]" "[]"
+      "{" "{}"
+      "}" "{}"
+      s))) ;; default
+
 
 (defn insert-result [^StyledTextArea codearea pe cmd]
   (let [caret-left? (= (.getCaretPosition codearea)
                        (-> codearea .getSelection .getStart))
         {:keys [length offset]} pe
-        delete? (#{:paredit-backward-delete :paredit-forward-delete} cmd)]
+        delete? (boolean (#{:paredit-backward-delete :paredit-forward-delete} cmd))]
 
     (when *debug* (println "caret-left?:" caret-left?))
     (when *debug* (println "delete?:" delete?))
@@ -58,7 +75,7 @@
     (doseq [{:keys [length offset text] :as mod} (:modifs pe)]
       (when *debug* (println "mod:" mod))
       (if (zero? length)
-        (.insertText codearea offset text)
+        (.insertText codearea offset  (correct-pairing codearea offset text))
         (.replaceText codearea offset (+ offset length) text)))
 
     ;; adjust caret and selection